fix(hosted): bind Linq group owner to adder
