+++ apps/web/.env.example
@@ -139,6 +139,10 @@ HOSTED_ONBOARDING_LINQ_CONVERSATION_PHONE_NUMBERS="+15550100001,+15550100002"
 HOSTED_ONBOARDING_LINQ_LOCAL_ALLOWED_INBOUND_PHONE_NUMBERS="+15550109999"
 # Optional soft cap for new home-line assignments. Once a line reaches this many active members, activation chooses another line in the pool.
 HOSTED_ONBOARDING_LINQ_MAX_ACTIVE_MEMBERS_PER_PHONE_NUMBER="1000"
+# Keep off for the compatibility deploy. Enable only after Linq confirms every
+# managed-line participant.added event includes added_by_handle; eligible first
+# group messages then retry until that event establishes the owner.
+HOSTED_LINQ_GROUP_OWNER_FROM_ADDER_REQUIRED="0"
 # Optional first-contact admission gate for unknown Linq/iMessage inbound join links.
 # Keep off by default. When enforced, unknown first contacts are classified before Murph sends an invite link.
 HOSTED_ONBOARDING_LINQ_FIRST_CONTACT_ADMISSION_MODE="off"
